update ni fgen example
