+++ examples/nifgen_standard_function/measurement.py
@@ -4,7 +4,7 @@ import contextlib
 import logging
 import pathlib
 import time
-from typing import Any, Dict, Tuple
+from typing import Tuple
 
 import click
 import grpc
@@ -13,12 +13,15 @@ import nifgen
 from _helpers import (
     ServiceOptions,
     configure_logging,
+    create_session_management_client,
+    get_grpc_device_channel,
     get_service_options,
     grpc_device_options,
     str_to_enum,
     use_simulation_option,
     verbosity_option,
 )
+from _nifgen_helpers import _create_nifgen_session
 
 import ni_measurementlink_service as nims
 
@@ -85,12 +88,7 @@ def measure(
 
     measurement_service.context.add_cancel_callback(cancel_callback)
 
-    session_management_client = nims.session_management.Client(
-        grpc_channel=measurement_service.get_channel(
-            provided_interface=nims.session_management.GRPC_SERVICE_INTERFACE_NAME,
-            service_class=nims.session_management.GRPC_SERVICE_CLASS,
-        )
-    )
+    session_management_client = create_session_management_client(measurement_service)
 
     with contextlib.ExitStack() as stack:
         reservation = stack.enter_context(
@@ -105,8 +103,9 @@ def measure(
             )
         )
 
+        grpc_device_channel = get_grpc_device_channel(measurement_service, nifgen)
         sessions = [
-            stack.enter_context(_create_nifgen_session(session_info))
+            stack.enter_context(_create_nifgen_session(session_info, grpc_device_channel))
             for session_info in reservation.session_info
         ]
 
@@ -163,41 +162,6 @@ def measure(
                         raise
 
     return ()
-
-
-def _create_nifgen_session(
-    session_info: nims.session_management.SessionInformation,
-) -> nifgen.Session:
-    options: Dict[str, Any] = {}
-    if service_options.use_simulation:
-        options["simulate"] = True
-        options["driver_setup"] = {"Model": "5423 (2CH)"}
-
-    session_kwargs: Dict[str, Any] = {}
-    if service_options.use_grpc_device:
-        session_grpc_address = service_options.grpc_device_address
-
-        if not session_grpc_address:
-            session_grpc_channel = measurement_service.get_channel(
-                provided_interface=nifgen.GRPC_SERVICE_INTERFACE_NAME,
-                service_class="ni.measurementlink.v1.grpcdeviceserver",
-            )
-        else:
-            session_grpc_channel = measurement_service.channel_pool.get_channel(
-                target=session_grpc_address
-            )
-        # Assumption: the pin map specifies one NI-FGEN session per instrument. If the pin map
-        # specified an NI-FGEN session per channel, the session name would need to include the
-        # channel name(s).
-        session_kwargs["grpc_options"] = nifgen.GrpcSessionOptions(
-            session_grpc_channel,
-            session_name=session_info.session_name,
-            initialization_behavior=nifgen.SessionInitializationBehavior.AUTO,
-        )
-
-    return nifgen.Session(
-        session_info.resource_name, session_info.channel_list, options=options, **session_kwargs
-    )
 
 
 @click.command
